+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 - [1.1.1 - 2024-04-24](#111---2024-04-24)
 - [1.2.0 - 2024-04-30](#120---2024-04-30)
 - [1.3.0 - 2024-05-08](#130---2024-05-08)
+- [1.3.1 - TBD](#131---TBD)
 
 ## TODO
 
-- [ ] Command history
+- [ ] better prompt format parsing
 
 ## [1.0.0](https://github.com/jtompkin/goclacker/releases/tag/v1.0.0) - 2024-04-19
 
@@ -70,3 +71,11 @@ Terminal woes
 ### Changed
 
 - Interactive mode generally better
+
+## [1.3.1](https://github.com/jtompkin/goclacker/releases/tag/v1.3.1) - TBD
+
+Topper
+
+### Changed
+
+- `&t` now `&Nt` format specifier: display top N values of the stack. Probably buggy